// qin_soul/src/qin-type.ts
export enum Nature {
    Bool = "Bool",
    Bit = "Bit",
    Byte = "Byte",
    Tiny = "Tiny",
    Small = "Small",
    Int = "Int",
    Long = "Long",
    BigInt = "BigInt",
    Serial = "Serial",
    BigSerial = "BigSerial",
    Float = "Float",
    Real = "Real",
    Double = "Double",
    Numeric = "Numeric",
    BigNumeric = "BigNumeric",
    Char = "Char",
    Chars = "Chars",
    Date = "Date",
    Time = "Time",
    DateTime = "DateTime",
    ZoneTime = "ZoneTime",
    Timestamp = "Timestamp",
    Bytes = "Bytes",
    Blob = "Blob",
    Text = "Text",
    Object = "Object"
}